pyproject.toml
[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[project]
name = "pyopengl-occlusion-test"
version = "0.1.0"
description = "OpenGL occlusion detection test using depth buffer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyOpenGL>=3.1.7",
    "numpy>=1.20.0",
    "PySide6>=6.0.0",
]

[project.scripts]
occlusion-test = "occlusion_test:main"

[tool.uv]
dev-dependencies = [
    "ruff>=0.15.0",
]

[tool.ruff]
